+++ src/smartcontracts/smartContractResults.dev.net.spec.ts
@@ -1,17 +1,17 @@
-import { Account } from "../account";
 import { NetworkConfig } from "../networkConfig";
-import { getDevnetProvider, loadContractCode, TestWallets } from "../testutils";
+import { loadContractCode, loadTestWallets, TestWallet } from "../testutils";
 import { TransactionWatcher } from "../transactionWatcher";
 import { ContractFunction, SmartContract } from ".";
 import { GasLimit } from "../networkParams";
 import { assert } from "chai";
+import { chooseProvider } from "../interactive";
 
 describe("fetch transactions from devnet", function () {
-    let devnet = getDevnetProvider();
-    let wallets = new TestWallets();
-    let aliceWallet = wallets.alice;
-    let alice = new Account(aliceWallet.address);
-    let aliceSigner = aliceWallet.signer;
+    let devnet = chooseProvider("local-testnet");;
+    let alice: TestWallet;
+    before(async function () {
+        ({ alice } = await loadTestWallets());
+    });
 
     it("counter smart contract", async function () {
         this.timeout(60000);
@@ -29,10 +29,10 @@ describe("fetch transactions from devnet", function () {
             gasLimit: new GasLimit(3000000)
         });
 
-        transactionDeploy.setNonce(alice.nonce);
-        await aliceSigner.sign(transactionDeploy);
+        transactionDeploy.setNonce(alice.account.nonce);
+        await alice.signer.sign(transactionDeploy);
 
-        alice.incrementNonce();
+        alice.account.incrementNonce();
 
         // ++
         let transactionIncrement = contract.call({
@@ -40,10 +40,10 @@ describe("fetch transactions from devnet", function () {
             gasLimit: new GasLimit(3000000)
         });
 
-        transactionIncrement.setNonce(alice.nonce);
-        await aliceSigner.sign(transactionIncrement);
+        transactionIncrement.setNonce(alice.account.nonce);
+        await alice.signer.sign(transactionIncrement);
 
-        alice.incrementNonce();
+        alice.account.incrementNonce();
 
         // Broadcast & execute
         await transactionDeploy.send(devnet);